requirements.txt
Flask
pymongo
PyJWT
bcrypt
random-password-generator
python-dotenv
